+++ src/common/errors.test.ts
@@ -1,11 +1,14 @@
 import {
   assertIsIOError,
+  BaseURLNotFoundError,
   ComponentUndefinedError,
+  CoverageFileNotFoundError,
   InstanceMissingError,
   MapUndefinedError,
   ProfileUndefinedError,
   ProviderJsonUndefinedError,
   RecordingPathUndefinedError,
+  RecordingsNotFoundError,
   SuperJsonNotFoundError,
   UnexpectedError,
 } from './errors';
@@ -223,6 +226,59 @@ describe('errors', () => {
     it('returns correct format', () => {
       expect(error.toString()).toEqual(
         'SuperJsonNotFoundError: No super.json found.'
+      );
+    });
+  });
+
+  describe('when throwing RecordingsNotFoundError', () => {
+    const error = new RecordingsNotFoundError('path/to/recording.json');
+
+    it('throws in correct format', () => {
+      expect(() => {
+        throw error;
+      }).toThrow(
+        'Recordings could not be found for running mocked tests at "path/to/recording.json".\nYou must call the live API first to record API traffic.\nUse the environment variable SUPERFACE_LIVE_API to call the API and record traffic.\nSee https://github.com/superfaceai/testing#recording to learn more.'
+      );
+    });
+
+    it('returns correct format', () => {
+      expect(error.toString()).toEqual(
+        'RecordingsNotFoundError: Recordings could not be found for running mocked tests at "path/to/recording.json".\nYou must call the live API first to record API traffic.\nUse the environment variable SUPERFACE_LIVE_API to call the API and record traffic.\nSee https://github.com/superfaceai/testing#recording to learn more.'
+      );
+    });
+  });
+
+  describe('when throwing BaseURLNotFoundError', () => {
+    const error = new BaseURLNotFoundError('provider');
+
+    it('throws in correct format', () => {
+      expect(() => {
+        throw error;
+      }).toThrow(
+        'No base URL was found for provider "provider", configure a service in provider.json.'
+      );
+    });
+
+    it('returns correct format', () => {
+      expect(error.toString()).toEqual(
+        'BaseURLNotFoundError: No base URL was found for provider "provider", configure a service in provider.json.'
+      );
+    });
+  });
+
+  describe('when throwing CoverageFileNotFoundError', () => {
+    const samplePath = 'path/to/coverage.json';
+    const error = new CoverageFileNotFoundError(samplePath);
+
+    it('throws in correct format', () => {
+      expect(() => {
+        throw error;
+      }).toThrow(`No coverage file at path "${samplePath}" found.`);
+    });
+
+    it('returns correct format', () => {
+      expect(error.toString()).toEqual(
+        `CoverageFileNotFoundError: No coverage file at path "${samplePath}" found.`
       );
     });
   });