+++ src/common/errors.ts
@@ -1,20 +1,25 @@
 import { SDKExecutionError } from '@superfaceai/one-sdk';
 import { inspect } from 'util';
 
-class ErrorBase extends Error {
-  constructor(public kind: string, public override message: string) {
+export class ErrorBase extends Error {
+  constructor(kind: string, message: string) {
     super(message);
-    this.name = kind;
 
     Object.setPrototypeOf(this, ErrorBase.prototype);
+
+    this.name = kind;
   }
 
-  get [Symbol.toStringTag](): string {
-    return this.kind;
+  public get [Symbol.toStringTag](): string {
+    return this.name;
   }
 
-  override toString(): string {
-    return `${this.kind}: ${this.message}`;
+  public get kind(): string {
+    return this.name;
+  }
+
+  public override toString(): string {
+    return `${this.name}: ${this.message}`;
   }
 }
 
@@ -107,10 +112,10 @@ export class SuperJsonLoadingFailedError extends ErrorBase {
 }
 
 export class RecordingsNotFoundError extends ErrorBase {
-  constructor() {
+  constructor(path: string) {
     super(
       'RecordingsNotFoundError',
-      'Recordings could not be found for running mocked tests.\nYou must call the live API first to record API traffic.\nUse the environment variable SUPERFACE_LIVE_API to call the API and record traffic.\nSee https://github.com/superfaceai/testing#recording to learn more.'
+      `Recordings could not be found for running mocked tests at "${path}".\nYou must call the live API first to record API traffic.\nUse the environment variable SUPERFACE_LIVE_API to call the API and record traffic.\nSee https://github.com/superfaceai/testing#recording to learn more.`
     );
   }
 }
@@ -120,6 +125,15 @@ export class BaseURLNotFoundError extends ErrorBase {
     super(
       'BaseURLNotFoundError',
       `No base URL was found for provider "${provider}", configure a service in provider.json.`
+    );
+  }
+}
+
+export class CoverageFileNotFoundError extends ErrorBase {
+  constructor(path: string) {
+    super(
+      'CoverageFileNotFoundError',
+      `No coverage file at path "${path}" found.`
     );
   }
 }