+++ src/common/io.ts
@@ -1,4 +1,5 @@
 import * as fs from 'fs';
+import { join as joinPath, resolve as resolvePath } from 'path';
 import rimrafCallback from 'rimraf';
 import { Writable } from 'stream';
 import { promisify } from 'util';
@@ -9,6 +10,8 @@ export const access = promisify(fs.access);
 export const mkdir = promisify(fs.mkdir);
 export const readFile = promisify(fs.readFile);
 export const rimraf = promisify(rimrafCallback);
+export const rename = promisify(fs.rename);
+export const readdir = promisify(fs.readdir);
 
 export interface WritingOptions {
   append?: boolean;
@@ -33,6 +36,26 @@ export async function exists(path: string): Promise<boolean> {
 
   // No error, no problem.
   return true;
+}
+
+/**
+ * Creates a directory without erroring if it already exists.
+ * Returns `true` if the directory was created.
+ */
+export async function mkdirQuiet(path: string): Promise<void> {
+  try {
+    await mkdir(path);
+  } catch (err: unknown) {
+    assertIsIOError(err);
+
+    // Allow `EEXIST` because scope directory already exists.
+    if (err.code === 'EEXIST') {
+      return;
+    }
+
+    // Rethrow other errors.
+    throw err;
+  }
 }
 
 /**
@@ -66,4 +89,22 @@ export function streamEnd(stream: Writable): Promise<void> {
     stream.once('close', resolve);
     stream.end();
   });
+}
+
+export async function readFilesInDir(path: string): Promise<string[]> {
+  const resolvedPath = resolvePath(path);
+  const dirents = await readdir(path, { withFileTypes: true });
+  const files: string[] = [];
+
+  for (const dirent of dirents) {
+    if (dirent.isDirectory()) {
+      files.push(
+        ...(await readFilesInDir(joinPath(resolvedPath, dirent.name)))
+      );
+    } else {
+      files.push(joinPath(resolvedPath, dirent.name));
+    }
+  }
+
+  return files;
 }